Add more abstractions for PickerFormCells for easier reusing
