+++ Buform/Platforms/Ios/Components/Picker/AsyncPickerFormCellBase.cs
@@ -3,14 +3,8 @@ using ObjCRuntime;
 namespace Buform;
 
 [Preserve(AllMembers = true)]
-public abstract class AsyncPickerFormCellBase<TAsyncPickerItem> : PickerFormCellBase<TAsyncPickerItem> where TAsyncPickerItem : class, IAsyncPickerFormItem
+public abstract class AsyncPickerFormCellBase<TAsyncPickerItem> : PresentedPickerFormCellBase<TAsyncPickerItem> where TAsyncPickerItem : class, IAsyncPickerFormItem
 {
-    protected virtual PickerPresenterBase<TAsyncPickerItem>? PickerPresenter
-    {
-        get;
-        private set;
-    }
-
     public override bool IsSelectable => !Item?.IsReadOnly ?? false;
 
     protected AsyncPickerFormCellBase()
@@ -24,88 +18,8 @@ public abstract class AsyncPickerFormCellBase<TAsyncPickerItem> : PickerFormCell
         /* Required constructor */
     }
 
-    protected virtual UIViewController CreateViewController(TAsyncPickerItem item)
+    protected override UIViewController CreateViewController(TAsyncPickerItem item)
     {
         return new AsyncPickerViewController(UITableViewStyle.InsetGrouped, item);
-    }
-
-    protected virtual void UpdateInputType()
-    {
-        if (Item == null)
-        {
-            return;
-        }
-
-        PickerPresenter?.Dispose();
-
-        PickerPresenter = Item.InputType switch
-        {
-            PickerInputType.Default
-                => new DefaultPickerPresenter<TAsyncPickerItem>(CreateViewController),
-            PickerInputType.Dialog
-                => new DialogPickerPresenter<TAsyncPickerItem>(CreateViewController),
-            PickerInputType.PopUp
-                => new DefaultPickerPresenter<TAsyncPickerItem>(CreateViewController),
-            _ => throw new ArgumentOutOfRangeException(nameof(Item.InputType), Item.InputType, null)
-        };
-    }
-
-    protected override void OnItemSet()
-    {
-        UpdateReadOnlyState();
-        UpdateLabel(Item?.Label);
-        UpdateInputType();
-        UpdateValue(Item?.FormattedValue);
-        UpdateValidationErrorMessage(Item?.ValidationErrorMessage);
-    }
-
-    protected override void OnItemPropertyChanged(string? propertyName)
-    {
-        switch (propertyName)
-        {
-            case nameof(Item.IsReadOnly):
-                UpdateReadOnlyState();
-                break;
-            case nameof(Item.Label):
-                UpdateLabel(Item?.Label);
-                break;
-            case nameof(Item.InputType):
-                UpdateInputType();
-                break;
-            case nameof(Item.FormattedValue):
-                UpdateValue(Item?.FormattedValue);
-                break;
-            case nameof(Item.ValidationErrorMessage):
-                UpdateValidationErrorMessage(Item?.ValidationErrorMessage);
-                break;
-        }
-    }
-
-    public override async void OnSelected()
-    {
-        base.OnSelected();
-
-        if (Item == null)
-        {
-            return;
-        }
-
-        if (PickerPresenter == null)
-        {
-            return;
-        }
-
-        await PickerPresenter.PickAsync(this, Item).ConfigureAwait(true);
-    }
-
-    protected override void Dispose(bool disposing)
-    {
-        if (disposing)
-        {
-            PickerPresenter?.Dispose();
-            PickerPresenter = null;
-        }
-
-        base.Dispose(disposing);
     }
 }
