+++ Buform/Platforms/Ios/Components/Picker/PopUpPickerFormCell.cs
@@ -1,0 +1,19 @@
+using ObjCRuntime;
+
+namespace Buform;
+
+[Preserve(AllMembers = true)]
+[Register(nameof(PopUpPickerFormCell))]
+public sealed class PopUpPickerFormCell : PopUpPickerFormCellBase<IPickerFormItem> 
+{
+    public PopUpPickerFormCell()
+    {
+        /* Required constructor */
+    }
+
+    public PopUpPickerFormCell(NativeHandle handle)
+        : base(handle)
+    {
+        /* Required constructor */
+    }
+}
